requirements.txt
# FastAPI and Server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0

# Environment Variables
python-dotenv>=1.0.0

# LangChain
langchain>=1.0.0
langchain-classic>=1.0.0
langchain-community>=0.4.0
langchain-core>=1.0.0
langchain-google-genai>=2.0.0
langchain-experimental>=0.4.0

# Database
SQLAlchemy>=2.0.0
